+++ src/main/java/com/jvnlee/catchdining/domain/payment/domain/PaymentType.java
@@ -1,4 +1,4 @@
-package com.jvnlee.catchdining.entity;
+package com.jvnlee.catchdining.domain.payment.domain;
 
 public enum PaymentType {
     CREDIT_CARD, NAVER_PAY, KAKAO_PAY
